fix(bindings): 4 more int32/int64 return-type mismatches
Surfaced by the un-skip of spanset.test and geoset.test in 729199c.
All are the same shape — function declared as returning
`LogicalType::INTEGER` (INT32) but the C implementation wrote
`int64_t` to the result Vector, which DuckDB 1.4 rejects at bind time
as "Expected INT64, found INT32".

- `SpatialSetFunctions::Set_value_n` (geoset valueN): read the second
  argument Vector as `int32_t` (not `int64_t`), matching the declared
  `LogicalType::INTEGER`.
- `SpansetFunctions::Spanset_num_spans`: `UnaryExecutor<string_t,int64_t>`
  → `<string_t,int32_t>`.
- `SpansetFunctions::Datespanset_num_dates`: same.
- `SpansetFunctions::Tstzspanset_num_timestamps`: same.

+++ src/geo/geoset.cpp
@@ -425,7 +425,8 @@ void SpatialSetFunctions::Set_value_n(DataChunk &args, ExpressionState &state, V
     n.Flatten(args.size());
 
     auto input_data = FlatVector::GetData<string_t>(input);
-    auto n_data = FlatVector::GetData<int64_t>(n);
+    // `valueN` is registered with second argument LogicalType::INTEGER.
+    auto n_data = FlatVector::GetData<int32_t>(n);
     auto result_data = FlatVector::GetData<string_t>(result);
 
     for (idx_t i = 0; i < args.size(); i++) {
